fix: deleteMovie bug

+++ controllers/movies.js
@@ -38,7 +38,8 @@ const addMovie = (req, res, next) => {
 
 const deleteMovie = (req, res, next) => {
   const movieId = req.params._id;
-  Movie.findOne({ movieId })
+  const owner = req.user._id;
+  Movie.findOne({ movieId, owner })
     .then((data) => {
       if (data === null) {
         throw new NotFoundError('Фильм не найден');
@@ -46,7 +47,7 @@ const deleteMovie = (req, res, next) => {
       if (data.owner._id.toString() !== req.user._id) {
         throw new ForbiddenError();
       }
-      return Movie.findOneAndRemove({ movieId })
+      return Movie.findOneAndRemove({ movieId, owner })
         .then((movie) => res.send({ message: `Фильм ${movie.nameRU} удален` }));
     })
     .catch(next);
